misc/HtmlTestSuite/results/test-1067.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Named items: img id &amp; name</title>
<link href="mailto:ms2ger@gmail.com" rel="author" title="Ms2ger"/>
<link href="https://html.spec.whatwg.org/multipage/#dom-document-nameditem" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><div id="log"/>
<div id="test">
<img id="a" name="b"/>
</div>
<script>
test(function() {
  assert_equals(document.a, document.getElementsByTagName("img")[0]);
  assert_equals(document['a'], document.getElementsByTagName("img")[0]);
  assert_equals(document.b, document.getElementsByTagName("img")[0]);
  assert_equals(document['b'], document.getElementsByTagName("img")[0]);
}, "img elements that have a name and id attribute, should be accessible by both values.");
</script>
</body></html>
